# Add join reordering