--- a/tests/queries/0_stateless/00826_cross_to_inner_join.sql
+++ b/tests/queries/0_stateless/00826_cross_to_inner_join.sql
@@ -1,6 +1,7 @@
 SET enable_optimize_predicate_expression = 0;
 SET optimize_move_to_prewhere = 1;
 SET convert_query_to_cnf = 0;
+SET reorder_joins = 0;
 
 select * from system.one l cross join system.one r;