… src/Interpreters/JoinReorderingVisitor.cpp
… src/Interpreters/JoinToSubqueryTransformVisitor.cpp
… src/Interpreters/TreeRewriter.cpp
… src/Interpreters/findGoodReordering.cpp
… src/Interpreters/getTableExpressions.cpp
… src/Interpreters/getTableExpressions.h
… tests/analyzer_tech_debt.txt
… tests/queries/0_stateless/00826_cross_to_inner_join_with_reordering.reference
… tests/queries/0_stateless/00826_cross_to_inner_join_with_reordering.sql
… tests/queries/0_stateless/02282_array_distance.sql
… utils/tune-join-coefficients/fit-regression.py
… utils/tune-join-coefficients/make-sql.py
… utils/tune-join-coefficients/run.sh