Automatically reorder joins.

The algorithm is quite simple as far as join reordering algorithms go.

The main search over join plans is a greedy algorithm, where the first two tables in the join order is brute-forced (i.e. it does a greedy search for every pair of starting tables). It has O(N^4) runtime, with a relatively small constant factor, so it's fast in practice.

The cost model comes from fitting a linear regression to predict the runtime of a two-table hash join.

The selectivity estimates are extremely naive (not based on any statistics).

Join keys are extracted from the WHERE clause and from ON clauses, when possible. If a join key is extracted from a WHERE clause, it is removed from the where clause.

A special optimization is done for queries like TPCH Q19. It has a join key inside an OR expression like `WHERE (t1.id = t2.id AND b) OR (t1.id = t2.id AND c) OR (t1.id = t2.id AND d)`. Since this can be rewritten to `WHERE (t1.id = t2.id) AND (b OR c OR d)`, it is valid to extract it as a join key. This allows a cross join to be turned into an inner join, which allows the query to complete in < 1 sec, instead of taking many days.

The transformation is done with an AST visitor. I realize there's a lot of work going on with the new analyzer framework, and it might better to do this transformation there. Still opted to do it like this, since the analyzer is still experimental and we needed something to speedup joins immediately. Hopefully the code can be a good starting point that can be incorporated into the new analyzer at a later stage.

Note: This was initially written against a branch of 23.3. There, every query in TPCH either got a speedup, or got the same plan as before. The benchmarks below are based on the current master branch. The benchmarks are noisy on my machine, but it looks like Q21 gets a slower query plan than before. I haven't looked into why this happens yet, but possibly the recent changes to joins (like more pushdown to right side of join) means that the cost model needs to be updated. I suspect it might be an easy fix, and I'll look into it, but decided to share the PR right away.


### Benchmarks:

I've benchmarked on TPCH, using queries adapted for CH from here: https://github.com/MonetDB/tpch-clickhouse

|TPCH sf=10|reorder_joins=0|reorder_joins=1|old/new (speedup)| |
|:----|:----|:----|:----|:----|
|Q1|0.398|0.398|1.00|no join in query|
|Q2|35.794|0.464|77.14| |
|Q3|3.001|0.724|4.15| |
|Q4|1.439|1.474|0.98|no join in query|
|Q5|3.761|0.801|4.70| |
|Q6|0.098|0.071|1.38|no join in query|
|Q7|14.567|1.983|7.35| |
|Q8|36.437|0.905|40.26| |
|Q9|213.599|2.266|94.26| |
|Q10|1.483|0.476|3.12| |
|Q11|0.168|0.142|1.18| |
|Q12|0.238|0.225|1.06| |
|Q13|2.896|2.8|1.03|LEFT JOIN in query, not reordered|
|Q14|0.208|0.201|1.03| |
|Q15|0.197|0.241|0.82| |
|Q16|0.112|0.117|0.96| |
|Q17|0.961|0.993|0.97| |
|Q18|4.167|1.345|3.10| |
|Q19|timeout|0.409|+inf| |
|Q20|0.835|0.896|0.93| |
|Q21|6.301|8.173|0.77| |
|Q22|0.231|0.289|0.80|no join in query|


|23.7| | | | |
|:----|:----|:----|:----|:----|
|TPCH sf=100|reorder_joins=0|reorder_joins=1| | |
|Q1|3.348|3.827|0.87|no join in query|
|Q2|3923.712|4.615|850.21| |
|Q3|36.158|9.893|3.65| |
|Q4|17.958|16.693|1.08|no join in query|
|Q5|51.63|7.497|6.89| |
|Q6|0.595|0.519|1.15|no join in query|
|Q7|92.788|20.844|4.45| |
|Q8|2161.383|9.789|220.80| |
|Q9|17643.531|33.278|530.19| |
|Q10|18.349|4.708|3.90| |
|Q11|1.675|1.315|1.27| |
|Q12|2.077|1.897|1.09| |
|Q13|34.952|33.946|1.03|LEFT JOIN in query, not reordered|
|Q14|2.345|2.304|1.02| |
|Q15|1.593|1.425|1.12| |
|Q16|1.002|0.899|1.11| |
|Q17|14.209|13.443|1.06| |
|Q18|61.426|19.589|3.14| |
|Q19|>850139|3.317|>256297.00| |
|Q20|10.001|9.257|1.08| |
|Q21|78.071|102.814|0.76| |
|Q22|3.977|4.107|0.97|no join in query|


(A note required by the TPC license: These are results from my machine, and are not official "TPC Benchmark Results", and not comparable to these.)


### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):

Automatically reorder joins to produce a faster query plan.