@larspars 2023-08-04
> It'd be great to see the benchmarking script in the PR for benchmark reproduction.

The benchmark was done in two steps:

1) Generate the data, using the code from tpc.org
2) Pipe in the queries from https://github.com/MonetDB/tpch-clickhouse , passing `--time` to clickhouse-client

The code from tpc.org is not redistributable. The queries from MonetDB are MPL 2.0 licensed, not sure if they can be included. I'm not sure if it's feasible to make a ready to run script to reproduce the numbers, but maybe a text file explaining how to reproduce them?