@larspars 2023-08-23
@vdimir @novikd 

There's an annoying issue relating to column names, and I need some help.

The current behaviour is as follows:
1) When there's one table, short names are used:
    ```sql
    select t1.id from t1
    ┌─id─┐
    ```

2) When there's two tables, qualified names are used for columns in the right table, iff the column name exists in the left table:
    ```sql
    select t1.id, t2.id from t1, t2
    ┌─id─┬─t2.id─┐
    ```

3) However, when there's more than two tables, qualified names are used for all columns, even the left table:
    ```sql
    select t1.id, t2.id, t3.id from t1, t2, t3
    ┌─t1.id─┬─t2.id─┬─t3.id─┐
    ```

For case 2, when there's two tables, it's possible to get different output when reorder_joins=1. Reordering can change which table is put on the left side, which affects column names when there's name clashes. I.e. the following is possible when reorder_joins=1:
```sql
select t1.id, t2.id from t1, t2
┌─t1.id─┬─id─┐
```
This can cause breakage if column names are used in outer queries:

```sql
select id from (select t1.id, t2.id from t1, t2)
```

I've tried to do the following to keep the old behaviour for 2 table queries:

1) Use `IdentifierSemantic::setColumnShortName` / `IdentifierSemantic::setColumnLongName` in JoinReorderingVisitor to set long and short column names according to their position before reordering. 

Result: It is overridden by `TranslateQualifiedNamesVisitor` called from `TreeRewriter::analyzeSelect(...)`, so it has no effect.

2) Don't use `TranslateQualifiedNamesVisitor` in `TreeRewriter::analyzeSelect(...)`.

Result: Breakage in later stages. As I remember, there's an error in `TreeRewriterResult::collectUsedColumns()`, because there's an expectation that columns from the left table use short names.

3) Add aliases for columns in the two-table case. I.e. convert `select t1.id, t2.id from t1, t2` to `select t1.id as id, t2.id as \`t2.id\` from t1, t2`.

Result: This seems to kill other optimizations. I saw that simply adding aliases resulted in much slower execution speed on some queries. It's also seems like a very brittle solution since you need to prove that aliases can be added without affecting the semantics of the query.

Any thoughts on this? It's really frustrating. I turned off reorder_joins for tests `01103_distributed_product_mode_local_column_renames` and `01824_prefer_global_in_and_join because of this issue`. Is it acceptable to merge with this problem existing? Any ideas on how to approach fixing it? Can we rename columns after execution?