@UnamedRus 2023-08-24
AFAIK, column names which ClickHouse generate is also not consistent with ANSI SQL or other DBMS, may be it's good time to fix that.

> This seems to kill other optimizations.

JOIN optimizations in ClickHouse quite fragile, so it's expected, may be you should aim to make it work with experimental analyzer first.