@larspars 2023-08-25
I worked around the column name issue by simply storing what the names would be before reordering, and applying those names again later. It meant that TreeRewriter::analyzeSelect(..) and SelectQueryExpressionAnalyzer::appendProjectResult(..) had to be updated to be aware of this, but it's still much less complex than any other approach I tried.