@Slach 2023-11-14
@alanpaulkwan check binary release ok
click to details
![image](https://github.com/ClickHouse/ClickHouse/assets/105560/00e8b0ee-2aba-40dc-a895-4c42bf268117)

click to summary and scroll down to artifacts
![image](https://github.com/ClickHouse/ClickHouse/assets/105560/64053a69-e933-4fcb-9d12-c9227528b653)

download selected artifacts, unzip it and look  and look to the URLS inside json

for example server packages url for this pr are depends on your platform


https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-common-static-23.9.1.40.x86_64.rpm
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-common-static-dbg-23.9.1.40.x86_64.rpm
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-common-static-dbg_23.9.1.40_amd64.deb
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-common-static-dbg_23.9.1.40_x86_64.apk
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-common-static_23.9.1.40_amd64.deb
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-common-static_23.9.1.40_x86_64.apk
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-server-23.9.1.40-amd64.tgz
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-server-23.9.1.40-amd64.tgz.sha512
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-server-23.9.1.40.x86_64.rpm
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-server_23.9.1.40_amd64.deb
https://s3.amazonaws.com/clickhouse-builds/PRs/52999/47bc3290a3035865b4e37261862fde3190b354f6/package_release/clickhouse-server_23.9.1.40_x86_64.apk


download common-static deb and server .deb for example and install it via `dpkg -i *.deb`